refactor(actor): move Actor/Singleton/HandlesKind/Dispatch back to aether-actor (issue 545 PR E4)
Pre-PR-C of issue 533 these markers lived in `aether-actor`. PR C
moved them down to `aether-data` to break a cycle the facade pattern
introduced (`aether-actor → aether-kinds → aether-actor`); marked
stopgap. PR E1 of issue 545 collapsed the facade pattern back out of
`aether-kinds` — the cycle that forced the down-move evaporated.

This PR restores the markers to their natural home alongside the rest
of the actor SDK.

- `aether-actor/src/actor.rs` regains the trait definitions it owned
  pre-PR-C (`Actor`, `Singleton`, `HandlesKind`) plus the native
  `Dispatch` trait the `#[actor]` proc-macro emits on chassis-cap
  inherent impls. `aether-data/src/actor.rs` is gone.
- `aether-data-derive` macro emits `::aether_actor::HandlesKind` /
  `::aether_actor::Dispatch` instead of `::aether_data::*`. No path
  changes for `Kind` / `Schema` / `ReplyTo` — the `Kind` derive stays
  rooted in aether-data; `ReplyTo` (the substrate-side dispatch type)
  also stays in aether-data so the move avoids a name clash with
  aether-actor's wasm-side `ReplyTo` (a u32 FFI handle, distinct
  shape).
- aether-substrate cap files (`audio.rs`, `handle.rs`, `io.rs`,
  `log.rs`, `net.rs`, `capability.rs`, `chassis_builder.rs`,
  `boot.rs`) switch their imports from `aether_data::{Actor,
  Singleton, Dispatch}` to `aether_actor::*`.

Net diff: +111/-118 across 14 files.

Out of scope: ADR superseding ADR-0075 (PR E5).
