+++ crates/aether-data-derive/src/lib.rs
@@ -1102,7 +1102,7 @@ fn expand_native_actor(item: ItemImpl) -> syn::Result<TokenStream2> {
     let handles_kind_impls = handlers.iter().map(|h| {
         let kind_ty = &h.kind_ty;
         quote! {
-            impl #impl_generics ::aether_data::HandlesKind<#kind_ty>
+            impl #impl_generics ::aether_actor::HandlesKind<#kind_ty>
                 for #self_ty #where_clause {}
         }
     });
@@ -1162,7 +1162,7 @@ fn expand_native_actor(item: ItemImpl) -> syn::Result<TokenStream2> {
     Ok(quote! {
         #(#handles_kind_impls)*
 
-        impl #impl_generics ::aether_data::Dispatch for #self_ty #where_clause {
+        impl #impl_generics ::aether_actor::Dispatch for #self_ty #where_clause {
             fn __dispatch(
                 &mut self,
                 __sender: ::aether_data::ReplyTo,
